requirements.txt
pyTelegramBotAPI~=4.22.1
django~=5.1.6
python-dotenv~=1.0.1
psycopg2-binary~=2.9.1
django-admin-sortable2~=2.2.3

#django-storages[s3]~=1.14.4
#boto3~=1.35.49
#locust~=2.32.1